chore: move src/form.tsx to src/components/Form.tsx
